queue_handler/requirements.txt
aio_pika==9.0.5
boto3==1.26.123
loguru==0.7.0
Pillow==9.5.0
pyarrow==11.0.0
sentry-sdk==1.20.0
uvloop==0.17.0
